Fix INSTALLDIRS setting.
It used to be that the search order of `@INC` was perl, then site. If you
installed a new version of something from CPAN, it would be installed into
site, by default. That meant that if it was in core, you'd never see the new
version, because the core code was in the "perl" lib.

To work around this, libraries found in both core and CPAN overrode their
install location to go into perl lib. This was a lousy situation. Now it's
over! Perl now looks in site first.